+++ fcnDVECORNER2PARAM.m
@@ -106,7 +106,7 @@ verticeList = [P1;P2;teRightProj;teLeftProj];
 % this ensures vertices from different vehicle/type(rotor vs wing) 
 % will not get WELDED
 try
-verticeList = [verticeList,reshape(repmat(vecDVEWING,1,4),[],1)];
+verticeList = [verticeList,double(reshape(repmat(vecDVEWING,1,4),[],1))];
 catch
 end
